+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Toast/UX/User Response/RichTextBox/BasicNotificationWithUserResponse.resx
