+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Toast/UX/User Response/Wrapped Label/BasicNotificationWithUserResponseWrappedLabel.resx
